+++ src/main/java/com/codecrafter/mahalaxmisandwich/services/impl/ToppingServiceImpl.java
@@ -1,5 +1,6 @@
 package com.codecrafter.mahalaxmisandwich.services.impl;
 
+import com.codecrafter.mahalaxmisandwich.entities.Item;
 import com.codecrafter.mahalaxmisandwich.entities.Topping;
 import com.codecrafter.mahalaxmisandwich.repositories.IToppingRepository;
 import com.codecrafter.mahalaxmisandwich.services.IToppingService;
@@ -7,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ToppingServiceImpl  implements IToppingService {
@@ -29,8 +31,16 @@ public class ToppingServiceImpl  implements IToppingService {
     }
 
     @Override
-    public void deleteToppingById(Long toppingId) {
-        toppingRepository.deleteById(toppingId);
+    public Topping deleteToppingById(Long toppingId) {
+        Optional<Topping> toppingOptional = toppingRepository.findById(toppingId);
+
+        if (toppingOptional.isPresent()) {
+            Topping topping = toppingOptional.get();
+            toppingRepository.deleteById(toppingId);
+            return topping;
+        } else {
+            throw new RuntimeException("Topping not found with id: " + toppingId);
+        }
     }
 
     @Override
